requirements.txt
asgiref==3.4.1
Django==3.2.8
django-debug-toolbar==3.2.2
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
Pillow==8.3.2
PyJWT==2.2.0
pytz==2021.3
sqlparse==0.4.2
